Module "party": Added namespace to class Party and removed dead code
- Added namespace LanSuite\Module\Party to class Party
- Added PHPdocs
- Reformatted code
- Removed dead code
- Removed various methods of class Party that were not in use
